requirements.txt
streamlit>=1.24.0
psycopg2-binary>=2.9.6
argon2-cffi>=21.3.0
python-dotenv>=1.0.0
pyjwt>=2.7.0
python-jose>=3.3.0
cryptography>=41.0.1
secure>=0.3.0
fastapi>=0.104.1
uvicorn>=0.24.0
websockets>=12.0
asyncpg>=0.29.0
pandas>=2.1.3
plotly>=5.18.0
streamlit-autorefresh>=0.0.1
requests>=2.31.0
